backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
